Add slide-out search sidebar to Navbar.vue
Code has been added to the Navbar.vue file to implement a slide-out search sidebar. This sidebar includes a search box and the layout for search results. It can be opened and closed by clicking on a button included in the NavBar. The slide-out sidebar offers an effective way to integrate search functionality without cluttering the user interface.