Implement loading screen for theme selection
The loading state and frontend have been adjusted to include a loading screen when selecting a theme. The relevant changes also involved updating the readme task list and a new test page to check for the functionality of the loading spinner. Selected theme is now also saved with each change instead of manually saving it.