Add Username component to prelobby page
Added a Username component to the prelobby page that prompts user to enter their username. This component is validated with vuelidate and opens on page load. The username state is managed globally using vuex in the lobby store. Minor changes in other components and files to accommodate the new Username component are also implemented.