Refactor code and modify user interface elements
This commit includes changes to several .vue files in the frontend components. It involves removal of some TODO comments, modifications in the HTML structure and CSS properties to improve the appearance and functionality of UI elements such as modals and input fields. Also, refactored the function for setting lobby details for better legibility and easier debugging.